fix(delete-ingredient): fix update view when ingredient deleted
